// src/notification/notification.enum.ts
export enum NotificationTypeEnum {
  PUBLIC = 'PUBLIC',
  NEW_CONTACT_MESSAGE = 'NEW_CONTACT_MESSAGE',
  AGORA_ACCESS_TOKEN_GENERATED = 'AGORA_ACCESS_TOKEN_GENERATED',
  AGORA_MEDIA_REQUEST_REPLIED = 'AGORA_MEDIA_REQUEST_REPLIED',
  AGORA_MEDIA_STATUS_CHANGED = 'AGORA_MEDIA_STATUS_CHANGED',
  LECTURER_APPROVAL = 'LECTURER_APPROVAL',
  PUBLICATION_APPROVAL = 'PUBLICATION_APPROVAL',
  PUBLICATION_REJECTION = 'PUBLICATION_REJECTION',
  MONTHLY_PAYOUT_COMPLETED = 'MONTHLY_PAYOUT_COMPLETED',
  ADDED_TO_DIPLOMA = 'ADDED_TO_DIPLOMA',
  POST_PUBLISHED = 'POST_PUBLISHED',
  ENCORAGEMENT_TO_PUBLISH_PROGRAM = 'ENCORAGEMENT_TO_PUBLISH_PROGRAM',
  //user notifications
  PURCHASE_CONFIRMATION = 'PURCHASE_CONFIRMATION',
  SPECIAL_DISCOUNT_OFFER = 'SPECIAL_DISCOUNT_OFFER',

  NEW_LEARNING_PROGRAM_AVAILABLE = 'NEW_LEARNING_PROGRAM_AVAILABLE',
  NEW_COURSE_AVAILABLE = 'NEW_COURSE_AVAILABLE',
  NEW_DIPLOMA_AVAILABLE = 'NEW_DIPLOMA_AVAILABLE',
  NEW_WORKSHOP_AVAILABLE = 'NEW_WORKSHOP_AVAILABLE',

  ENCORAGEMENT_TO_COMPLETE_PROGRAM = 'ENCORAGEMENT_TO_COMPLETE_PROGRAM',
  ENCORAGEMENT_TO_COMPLETE_COURSE = 'ENCORAGEMENT_TO_COMPLETE_COURSE',
  ENCORAGEMENT_TO_COMPLETE_DIPLOMA = 'ENCORAGEMENT_TO_COMPLETE_DIPLOMA',
  ENCORAGEMENT_TO_COMPLETE_WORKSHOP = 'ENCORAGEMENT_TO_COMPLETE_WORKSHOP',

  PROGRAM_COMPLETED_CERTIFICATE_AVAILABLE = 'PROGRAM_COMPLETED_CERTIFICATE_AVAILABLE',
  COURSE_COMPLETED_CERTIFICATE_AVAILABLE = 'COURSE_COMPLETED_CERTIFICATE_AVAILABLE',
  WORKSHOP_COMPLETED_CERTIFICATE_AVAILABLE = 'WORKSHOP_COMPLETED_CERTIFICATE_AVAILABLE',
  DIPLOMA_COMPLETED_CERTIFICATE_AVAILABLE = 'DIPLOMA_COMPLETED_CERTIFICATE_AVAILABLE',

  USER_ASSIGNED_TO_PROGRAM = 'USER_ASSIGNED_TO_PROGRAM',
  USER_ASSIGNED_TO_DIPLOMA = 'USER_ASSIGNED_TO_DIPLOMA',
  USER_ASSIGNED_TO_COURSE = 'USER_ASSIGNED_TO_COURSE',
  USER_ASSIGNED_TO_WORKSHOP = 'USER_ASSIGNED_TO_WORKSHOP',

  //common
  CONTACT_US_TICKET_CONFIRMATION = 'CONTACT_US_TICKET_CONFIRMATION'
}
// ...
export enum SiteNotificationsTypeEnum {
  //lecturer notifications
  LECTURER_APPROVAL = 'LECTURER_APPROVAL',
  PUBLICATION_APPROVAL = 'PUBLICATION_APPROVAL',
  PUBLICATION_REJECTION = 'PUBLICATION_REJECTION',
  MONTHLY_PAYOUT_COMPLETED = 'MONTHLY_PAYOUT_COMPLETED',
  ADDED_TO_DIPLOMA = 'ADDED_TO_DIPLOMA',
  POST_PUBLISHED = 'POST_PUBLISHED',
  ENCORAGEMENT_TO_PUBLISH_PROGRAM = 'ENCORAGEMENT_TO_PUBLISH_PROGRAM',
  //user notifications
  PURCHASE_CONFIRMATION = 'PURCHASE_CONFIRMATION',
  SPECIAL_DISCOUNT_OFFER = 'SPECIAL_DISCOUNT_OFFER',
  NEW_PROGRAM_AVAILABLE = 'NEW_LEARNING_PROGRAM_AVAILABLE',
  ENCORAGEMENT_TO_COMPLETE_PROGRAM = 'ENCORAGEMENT_TO_COMPLETE_PROGRAM',
  PROGRAM_COMPLETED_CERTIFICATE_AVAILABLE = 'PROGRAM_COMPLETED_CERTIFICATE_AVAILABLE',
  USER_ASSIGNED_TO_PROGRAM = 'USER_ASSIGNED_TO_PROGRAM',

  //common
  CONTACT_US_TICKET_CONFIRMATION = 'CONTACT_US_TICKET_CONFIRMATION'
}

registerEnumType(SiteNotificationsTypeEnum, {
  name: 'SiteNotificationsTypeEnum'
});

export enum BoardNotificationSortEnum {
  TIMES_SENT = 'timesSent',
  CREATED_AT = 'createdAt'
}
registerEnumType(BoardNotificationSortEnum, {
  name: 'BoardNotificationSortEnum'
});

export enum NotificationParentTypeEnum {
  COURSE = 'COURSE',
  WORKSHOP = 'WORKSHOP',
  DIPLOMA = 'DIPLOMA',
  CERTIFICATE = 'CERTIFICATE'
}